server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "databases>=0.9.0",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-dotenv>=0.5.2",
    "pytest-sugar>=1.0.0",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]

[tool.uv.workspace]
members = ["fastapi"]
